일정 생성 성공 후 호출할 일정 목록 페이지 수정

+++ Front/src/pages/mypage/PlanList.js
@@ -5,15 +5,16 @@ import { Link, useNavigate, useParams } from "react-router-dom";
 function PlanList() {
     const [scheduleList, setSchduleList] = useState([]);
 
-    const params = useParams();
     const navigate = useNavigate();
     
     useEffect(() => {
         async function getPlanList() {
+            
             try {
-                const result = await axios.get("http://localhost:8080/mypage/planlist");
+                const result = await axios.get(`http://localhost:8080/mypage/planlist/${window.localStorage.getItem("userid")}`);
                 console.log(result);
                 setSchduleList(result.data);
+                console.log(scheduleList);
             } catch (error) {
                 console.log(error);
             }
@@ -21,15 +22,19 @@ function PlanList() {
         getPlanList();
     }, [])
 
-    console.log(scheduleList.map((schedule) => schedule.schedule_id));
+    // console.log(scheduleList.map((schedule) => schedule.schedule_id));
     
     const deleteClick = async (scheduleId) => {
-        try {
-            await axios.delete(`http://localhost:8080/mypage/planlist/plan-delete/${scheduleId}`);
-            alert("삭제 되었습니다");
-            navigate("/mypage/planlist");
-        } catch (error) {
-            alert("네트워크 문제로 삭제가 되지 않았습니다.");
+        if (window.confirm("삭제하시겠습니까?")) {
+            console.log(scheduleId);
+            try {
+                await axios.delete(`http://localhost:8080/mypage/planlist/plan-delete/${scheduleId}`);
+                alert("삭제 되었습니다.");
+                window.location.reload();
+                navigate("/mypage/planlist");
+            } catch (error) {
+                alert("네트워크 문제로 삭제가 되지 않았습니다.");
+            }
         }
     }
     
@@ -47,20 +52,11 @@ function PlanList() {
                 </thead>
                 <tbody>
                     {scheduleList.map((schedule, index) => {
-                        // const deleteClick = async (scheduleId) => {
-                        //     try {
-                        //         await axios.delete(`http://localhost:8080/mypage/planlist/plan-delete/${scheduleId}`);
-                        //         alert("삭제 되었습니다.");
-                        //         navigate("/mypage/planlist");
-                        //     } catch (error) {
-                        //         alert("네트워크 문제로 삭제가 되지 않았습니다.");
-                        //     }
-                        // }
                         return (
                             <tr key={index}>
                                 <td>{index+1}</td>
                                 <td>
-                                    <Link to={`/mypage/planlist/${schedule.schedule_id}`} style={{textDecoration:'none'}}>{schedule.title}</Link>
+                                    <Link to={`/mypage/planlist/plan/${schedule.schedule_id}`} style={{textDecoration:'none'}}>{schedule.title}</Link>
                                 </td>
                                 <td>
                                     {schedule.startDate} - {schedule.endDate}
@@ -69,27 +65,13 @@ function PlanList() {
                                 <div>
                                     <td>
                                         <div style={{display:'flex', justifyContent:'center'}}>
-                                            <button type="button" className="btn btn-outline-secondary btn-sm" data-bs-toggle="modal" data-bs-target="#exampleModal">삭제</button>
+                                            <button 
+                                                onClick={() => deleteClick(schedule.schedule_id)}
+                                                type="button" className="btn btn-outline-secondary btn-sm">
+                                                삭제
+                                            </button>
                                         </div>
                                     </td>
-
-                                    <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-                                        <div class="modal-dialog modal-dialog-centered modal-sm">
-                                        <div class="modal-content">
-                                            <div class="modal-header">
-                                                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                                            </div>
-
-                                            <div class="modal-body">삭제하시겠습니까?</div>
-
-                                            <div class="modal-footer">
-                                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">취소</button>
-                                            <button onClick={() => deleteClick(schedule.schedule_id)} class="btn btn-primary">확인</button>
-                                            {/* 확인 누르면 DB에서 삭제되도록 */}
-                                            </div>
-                                        </div>
-                                        </div>
-                                    </div>
                                     </div>
                                 </div>
                             </tr>
@@ -97,7 +79,7 @@ function PlanList() {
                     })}
                 </tbody>
             </table>
-            </div>
+        </div>
     )
 }
